Add compatibility for Rails protect_from_forgery
Make auto-session-timeout compatible with Rails' protect_from_forgery feature by adding a condition to prevent the session from being timed out if the user is logging in.  This condition is necessary because the user will receive a 'CSRF token invalid' error if they try logging in after they have been on the login page for longer than the configured timeout value (i.e. the time specified by auto_session_timeout in application_controller.rb).

If a developer wishes to take advantage of this feature, they can simply add a parameter to the auto_session_timeout declaration in application_controller.rb.  The parameter indicates the relative path of the sign in page.  For instance:  auto_session_timeout 1.hour,  '/users/sign_in'