Add checks for compatibility with ruby v3.1 and v3.2 and Rails v7.1.
I found that some versions of Rails did not seem to work with Ruby v2.4 so I excluded them from the matrix. There is also an issue with Rails 7.0 and Ruby > v3.0 so I changed the Rails version it is tested against to 7.0.1 (which is the first version that fixed the issue).